Refactored InitialScreen Model and Controller
Changed how the Green and Red players were stored in the InitialScreenModel. Didn't make sense to have them be in 2-dimensional array anymore. Refactored the Controller to work with a one dimensional array instead of a two dimensional one as well. Also, added functionality to the buttons at the bottom of the screen. The Start Match button prints out all players entered into the Initial Screen form, and the Clear Inputs button clears all inputs in the form.